+++ config/initializers/presentable.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+ActiveSupport.on_load :active_model do
+	include Magic::Presentable
+end