+++ lib/magic/presenter.rb
@@ -6,6 +6,8 @@ require_relative 'presenter/version'
 require_relative 'presenter/engine'
 
 module Magic
+	autoload :Presentable, 'magic/presentable'
+
 	# Presentation layer for Rails models
 	module Presenter
 		autoload :Base, 'magic/presenter/base'