+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ### Added
 
 - `Magic::Presenter::Base` — a basic presenter class.
+- `Magic::Presentable` to be included in presentable classes.
+  - `#decorate`,
+  - `#decorate!`,
+  - `#decorated`,
+  - `#decorated?`.